+++ random_agent.py
@@ -7,8 +7,8 @@ LOGGER = logging.getLogger(__name__)
 env = gym.make('Phoenix-ram-v0')
 
 tot_rewards = []
-quiet = False
-for i_episode in range(2):
+quiet = True
+for i_episode in range(100):
     observation = env.reset()
     tot_reward = 0
     print '##################'
@@ -16,12 +16,8 @@ for i_episode in range(2):
         if not quiet:
             env.render()
         action = env.action_space.sample()
-        #print 'action: %s' % action
         observation2, reward, done, info = env.step(action)
         if reward:
-            #print observation
-            print info
-            print reward
             tot_reward += reward
         if done:
             print("Episode finished after {} timesteps".format(t+1))
